remove logconfig from KafkaMetadataLog
